Project_DIR/requirements.txt
bcrypt==3.1.4
cffi==1.11.5
dataclasses==0.6
dj-database-url==0.5.0
Django==2.0.6
django-crispy-forms==1.7.2
django-pagedown==1.0.5
gunicorn==19.8.1
Pillow==5.1.0
pkg-resources==0.0.0
psycopg2==2.7.4
pycparser==2.18
pytz==2018.4
six==1.11.0
sqlparse==0.2.4
whitenoise==3.3.1
